style(breadcrumbd): render breadcrumb and accept props
insetead of using higher order component on each page
